address feedback: break test-only bits of ParallelPrefetch out into a partial
